Rearrange tests file structure to be cleaner
