+++ tests/desmume/desmume_utils.py
