+++ tests/desmume/test_data/test_custom_dig_spot_items.dsv
