+++ tests/desmume/test_data/test_custom_shop_items.dsv
