+++ tox.ini
@@ -8,7 +8,7 @@ deps =
     pytest
     pytest-xdist
 commands =
-    pytest --numprocesses auto {posargs} --ignore=tests/test_desmume.py
+    pytest --numprocesses auto {posargs} --ignore=tests/desmume
 
 [testenv:test-desmume]
 passenv =
@@ -19,7 +19,7 @@ deps =
     pytest-xdist
     py-desmume>=0.0.4.post2
 commands =
-    pytest --numprocesses auto {posargs} tests/test_desmume.py
+    pytest --numprocesses auto {posargs} tests/desmume
 
 [pytest]
 addopts = --strict-markers --showlocals --verbose
